move action to common folder
